src/test/java/com/sab/lombokbasics/services/JuiceCsvServiceImplTest.java
package com.sab.lombokbasics.services;

import com.sab.lombokbasics.model.JuiceCSVRecord;
import org.junit.jupiter.api.Test;
import org.springframework.util.ResourceUtils;

import java.io.File;
import java.io.FileNotFoundException;
import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;

class JuiceCsvServiceImplTest {
    JuiceCsvService beerCsvService = new JuiceCsvServiceImpl();

    @Test
    void convertCSV() throws FileNotFoundException {

        File file = ResourceUtils.getFile("classpath:csvdata/juice.csv");

        List<JuiceCSVRecord> recs = beerCsvService.convertCSV(file);

        System.out.println(recs.size());

        assertThat(recs.size()).isGreaterThan(0);
    }
}